hugo: make minification keep quotes to work around go get
The go get HTML parser is not actually an HTML parser, but is the
encoding/xml code. This means that valid HTML containing unquoted
attributes fails parsing by go get when attempting to obtain go-import
and go-source redirections. The minifier provides the option to keep
quotes, so let's do that until go get is fixed.